Move generateSnippets functions to their corresponding files
