+++ packages/wesl-ext-typegpu/src/imports.mjs
@@ -3,16 +3,40 @@
 import { assertDefined } from './utils.mjs';
 
 /** @typedef {import("wesl").AbstractElem} AbstractElem */
+/** @typedef {import("wesl").StructElem} StructElem */
 /** @typedef {import("wesl").ImportElem} ImportElem */
 /** @typedef {import("wesl").ImportStatement} ImportStatement */
+/** @typedef {import("wesl").TypeRefElem} TypeRefElem */
 /** @typedef {{ path: string, finalSegment: string }} ImportInfo */
+
+/**
+ * @param {StructElem[]} structElems
+ * @param {ImportElem[]} importElems
+ * @param {Set<string>} importsNamespace
+ */
+export function generateImportSnippets(
+  structElems,
+  importElems,
+  importsNamespace,
+) {
+  const { inlinedImports, directImports } = findImportsUsedInStructs(
+    structElems,
+    importsNamespace,
+  );
+
+  const imports = [
+    `import * as d from 'typegpu/data'`,
+    ...parseImports(importElems, directImports, inlinedImports),
+  ];
+  return imports;
+}
 
 /**
  * @param {ImportElem[]} importElems
  * @param {Set<string>} identifiersToImport
  * @param {Set<string>} inlinedImports
  */
-export function parseImports(importElems, identifiersToImport, inlinedImports) {
+function parseImports(importElems, identifiersToImport, inlinedImports) {
   /** @type {Map<string, ImportInfo>} */
   const importOfAlias = generateImportMap(importElems);
 
@@ -105,4 +129,41 @@ function generateImportMap(importElems) {
  */
 function generateImport(path, item, alias) {
   return `import { ${item}${alias ? ` as ${alias}` : ''} } from '${path}.wesl?typegpu';`;
+}
+
+/**
+ * @param {StructElem[]} structElements
+ * @param {Set<string>} importsNamespace
+ */
+function findImportsUsedInStructs(structElements, importsNamespace) {
+  /** @type {Set<string>} */
+  const inlinedImports = new Set();
+  /** @type {Set<string>} */
+  const directImports = new Set();
+
+  /**
+   * @param {TypeRefElem} typeRef
+   */
+  function findUsedImportsInType(typeRef) {
+    const name = typeRef.name.originalName;
+    if (name.includes('::')) {
+      inlinedImports.add(name);
+    }
+    if (importsNamespace.has(name)) {
+      directImports.add(name);
+    }
+    for (const subtype of typeRef.templateParams ?? []) {
+      if (subtype.kind === 'type') {
+        findUsedImportsInType(subtype);
+      }
+    }
+  }
+
+  for (const struct of structElements) {
+    for (const member of struct.members) {
+      findUsedImportsInType(member.typeRef);
+    }
+  }
+
+  return { inlinedImports, directImports };
 }
